+++ CHANGELOG.rst
@@ -3,6 +3,12 @@
 CHANGELOG for ToDonePy
 ======================
 
+v3.0.0
+------
+- BREAKING: Add delimiter parameter to Filer class
+- BREAKING: Add headings to ``TODO.tsv``
+- Add itemsetter function
+
 v2.3.1
 ------
 - Update requirements